+++ ext/etl/runtime/all.go
@@ -12,9 +12,11 @@ import (
 )
 
 const (
-	Py38  = "python3.8v2"
+	Py39  = "python3.9v2"
 	Py310 = "python3.10v2"
 	Py311 = "python3.11v2"
+	Py312 = "python3.12v2"
+	Py313 = "python3.13v2"
 )
 
 type (
@@ -25,9 +27,11 @@ type (
 		DepsEnvName() string
 	}
 	runbase struct{}
-	py38    struct{ runbase }
+	py39    struct{ runbase }
 	py310   struct{ runbase }
 	py311   struct{ runbase }
+	py312   struct{ runbase }
+	py313   struct{ runbase }
 )
 
 var (
@@ -51,8 +55,8 @@ func GetNames() (names []string) {
 }
 
 func init() {
-	all = make(map[string]runtime, 3)
-	for _, r := range []runtime{py38{}, py310{}, py311{}} {
+	all = make(map[string]runtime, 5)
+	for _, r := range []runtime{py39{}, py310{}, py311{}, py312{}, py313{}} {
 		if _, ok := all[r.Name()]; ok {
 			debug.Assert(false, "duplicate type "+r.Name())
 		} else {
@@ -65,11 +69,17 @@ func (runbase) CodeEnvName() string { return "AISTORE_CODE" }
 func (runbase) DepsEnvName() string { return "AISTORE_DEPS" }
 
 // container images: "aistorage/runtime_python:<TAG>"
-func (py38) Name() string    { return Py38 }
-func (py38) PodSpec() string { return strings.ReplaceAll(pyPodSpec, "<TAG>", "3.8v2") }
+func (py39) Name() string    { return Py39 }
+func (py39) PodSpec() string { return strings.ReplaceAll(pyPodSpec, "<TAG>", "3.9v2") }
 
 func (py310) Name() string    { return Py310 }
 func (py310) PodSpec() string { return strings.ReplaceAll(pyPodSpec, "<TAG>", "3.10v2") }
 
 func (py311) Name() string    { return Py311 }
 func (py311) PodSpec() string { return strings.ReplaceAll(pyPodSpec, "<TAG>", "3.11v2") }
+
+func (py312) Name() string    { return Py312 }
+func (py312) PodSpec() string { return strings.ReplaceAll(pyPodSpec, "<TAG>", "3.12v2") }
+
+func (py313) Name() string    { return Py313 }
+func (py313) PodSpec() string { return strings.ReplaceAll(pyPodSpec, "<TAG>", "3.13v2") }